Improve docs, move Qurl class outside of template tag module.
- Present Qurl class with examples in README.
- Move Qurl class to a different module, expose it through top level module.
- Add docstrings to Qurl class.
