+++ qurl_templatetag/__init__.py
@@ -1,3 +1,5 @@
+from .qurl import Qurl  # noqa
+
 __author__ = """Sophilabs"""
 __email__ = 'hi@sophilabs.com'
 __version__ = '0.0.12'
